stock-analyst-app/requirements.txt
flask==2.3.3
boto3==1.34.0
requests==2.31.0
pymongo>=4.6.0
motor>=3.3.0
feedparser==6.0.11
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
python-dotenv==1.0.0
yfinance>=0.2.32
beautifulsoup4==4.12.2
lxml>=4.9.3
schedule==1.2.0
